inventory-worker/requirements.txt
tortoise-orm
asyncpg
pika
asyncio